fix: move build hook helper into library
