+++ hook/build.dart
@@ -4,7 +4,7 @@ import 'dart:io';
 
 import 'package:hooks/hooks.dart';
 
-import 'desktop_artifacts.dart';
+import '../lib/src/native/desktop_artifacts.dart';
 
 void main(List<String> args) async {
   await build(args, (input, output) async {
